google-group-sync/src/requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
requests==2.31.0
google-cloud-secret-manager==2.16.4
